test: de-duplicate some methods
