+++ tests/Dfe.PlanTech.Web.E2ETests/cypress/e2e/pages/question-page.cy.js
@@ -9,7 +9,7 @@ describe("Question page", () => {
 
     //Navigate to first question
     cy.clickContinueButton();
-    
+
     cy.injectAxe();
   });
