+++ tests/Dfe.PlanTech.Web.E2ETests/cypress/e2e/pages/recommendation-page.cy.js
@@ -1,14 +1,15 @@
-let selectedQuestionsWithAnswers = [];
-
 describe("Recommendation Page", () => {
     const url = "/self-assessment";
 
     beforeEach(() => {
-        selectedQuestionsWithAnswers = [];
-
         cy.loginWithEnv(url);
 
-        navigateToRecommendationPage();
+        cy.url().should("contain", "self-assessment");
+
+        cy.clickFirstSection();
+        cy.clickContinueButton();
+
+        cy.navigateToRecommendationPage();
 
         cy.url().should("contain", "recommendation");
 
@@ -36,42 +37,3 @@ describe("Recommendation Page", () => {
     });
 
 });
-
-const navigateToRecommendationPage = () => {
-    cy.clickFirstSection();
-    cy.clickContinueButton();
-
-    return navigateThroughCheckAnswersPage().then((res) => cy.wrap(res)).then(() => submitAnswers()).then((onSelfAssessmentPage) => {
-        if (!onSelfAssessmentPage) {
-            return Promise.resolve();
-        }
-
-        cy.get('a[href*="/recommendation/"]').first().click();
-    });
-};
-
-const navigateThroughCheckAnswersPage = () => {
-    return navigateThroughQuestions().then(() => cy.wrap(selectedQuestionsWithAnswers));
-};
-
-const navigateThroughQuestions = () => {
-    return cy
-        .get("main")
-        .then(($main) => $main.find("form div.govuk-radios").length > 0)
-        .then((onQuestionPage) => {
-            if (!onQuestionPage) {
-                return Promise.resolve();
-            }
-
-            cy.selectFirstRadioButton().then((questionWithAnswer) =>
-                selectedQuestionsWithAnswers.push(questionWithAnswer)
-            );
-            cy.saveAndContinue();
-
-            return navigateThroughQuestions();
-        })
-        .then(() => cy.wrap(selectedQuestionsWithAnswers));
-};
-
-const submitAnswers = () =>
-    cy.get("button.govuk-button").contains("Save and submit").click();
