+++ tests/Dfe.PlanTech.Web.E2ETests/cypress/support/commands.js
@@ -3,3 +3,4 @@ import "../commands/self-assessment-page";
 import "../commands/axe-tools";
 import "../commands/helpers";
 import "../commands/question-page";
+import "../commands/questionnaire-stages";
